Moved header and give a ride header to partials folder
